refactor(scaffold): templates as real files, not escaped TS strings
The files `init`/`add` scaffold lived as TS string literals in scaffold-templates.ts, so any
nested code had to escape its backticks/backslashes — `\\s+` in the word-count tool, every
`\`…\`` in the channel templates, and the whole telegram-send tool written with string
concatenation purely to dodge nested backticks. Template code was not real code: no
highlighting, no editor lint, escaping noise.

Move the content to real files under templates/ (AGENTS.md, house-style.md, tool.ts,
fastagent.config.mjs, gitignore, env.example, channels/{github,telegram}.ts,
tools/telegram-send.ts), read at scaffold time via `new URL("./templates/…", import.meta.url)`.
The exports keep their names (SKILL_MD, TOOL_TS, …) so init.ts/add-channel.ts are untouched;
only the parametric pieces (the AGENTS.md tool line, package.json, toPackageName) stay as code.

- templates/ is excluded from this package's tsc + biome (it imports the published
  @kid7st/fastagent, not this source) and copied into dist/ by the postbuild step; the
  package already ships dist, so npm pack includes it.
- Resolution works in both trees: src/ (vitest) and dist/ (the built CLI), since the loader
  resolves relative to its own module URL and the build colocates templates.
- telegram-send now reads as clean template literals instead of "a" + b concatenation — the
  generated tool is functionally identical, just legible. The de-escaped word-count regex is
  `/\s+/`, not `/\\s+/`.

Verified: init.test green; the built CLI scaffolds init + `add telegram` from dist with the
de-escaped content; npm pack lists all nine template files.
